requirements.txt
discord.py
mysql-connector-python
groq
langdetect
aiohttp
python-dotenv
loguru
fastapi
uvicorn
pydantic
PyJWT
requests
